@@ -5,7 +5,7 @@ repository: https://github.com/petercinibulk/envied
 homepage: https://github.com/petercinibulk/envied
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 # dependencies:
 #   path: ^1.8.0
