@@ -5,7 +5,7 @@ repository: https://github.com/petercinibulk/envied
 homepage: https://github.com/petercinibulk/envied
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
   envied:
@@ -19,10 +19,6 @@ dependencies:
   analyzer: ^2.1.0
 
 dev_dependencies:
-  lints: ^2.0.0
   test: ^1.16.0
   source_gen_test: ^1.0.3
 
-analyzer:
-  enable-experiment:
-    - non-nullable
